+++ pganonymizer/exceptions.py
@@ -3,7 +3,11 @@ class PgAnonymizeException(Exception):
 
 
 class InvalidFieldProvider(PgAnonymizeException):
-    """Raised if an unknown field provider was used."""
+    """Raised if an unknown field provider was used in the schema."""
+
+
+class InvalidProvider(PgAnonymizeException):
+    """Raised if an unknown provider class was requested."""
 
 
 class BadDataFormat(PgAnonymizeException):